Changed example to run ofMainLoop() everytime a helpBox is opened as without this the main app stops

+++ src/HelpBox.h
@@ -46,7 +46,7 @@ public:
 	float lineHeight;
 	float x;
 	float y;
-	float fontSize;
+	int fontSize;
 
 	int numLines;
 	bool wordsWereCropped;
